+++ packages/common/src/projects/_internal/ProjectUiSchemaSettings.ts
@@ -1,0 +1,40 @@
+import { IArcGISContext } from "../../ArcGISContext";
+import { IUiSchema } from "../../core/schemas/types";
+import { IHubProject } from "../../core/types";
+
+/**
+ * @private
+ * constructs the settings uiSchema for Hub Projects.
+ * This defines how the schema should be rendered
+ * in the project settings pane
+ */
+export const buildUiSchema = async (
+  i18nScope: string,
+  options: Partial<IHubProject>,
+  context: IArcGISContext
+): Promise<IUiSchema> => {
+  return {
+    type: "Layout",
+    elements: [
+      {
+        type: "Section",
+        labelKey: `${i18nScope}.sections.mapSettings.label`,
+        elements: [
+          {
+            type: "Control",
+            scope: "/properties/view/properties/mapSettings",
+            labelKey: `${i18nScope}.fields.mapSettings.label`,
+            options: {
+              type: "Control",
+              control: "hub-composite-input-map-settings",
+              // the settings that are visible for configuring the map
+              visibleSettings: ["gallery"],
+              // if the map preview is displayed
+              showPreview: true,
+            },
+          },
+        ],
+      },
+    ],
+  };
+};